changed the time when casting data type to float32.
for now, data conversation only occur when extracting experiences from the replay buffer or invoking a function with `@tf.function`.
see `tf2algos/base.py` -> `_cast` function for detail.
@kmakeev

+++ Algorithms/tf2algos/policy.py
@@ -100,13 +100,13 @@ class Policy(Base):
         if not self.action_type == 'continuous':
             a = sth.action_index2one_hot(a, self.a_dim_or_list)
         self.data = self.data.append({
-            's': s.astype(np.float32),
-            'visual_s': visual_s.astype(np.float32),
-            'a': a.astype(np.float32),
-            'r': r.astype(np.float32),
-            's_': s_.astype(np.float32),
-            'visual_s_': visual_s_.astype(np.float32),
-            'done': done.astype(np.float32)
+            's': s,
+            'visual_s': visual_s,
+            'a': a,
+            'r': r,
+            's_': s_,
+            'visual_s_': visual_s_,
+            'done': done
         }, ignore_index=True)
 
     def off_store(self, s, visual_s, a, r, s_, visual_s_, done):
@@ -119,13 +119,13 @@ class Policy(Base):
         if not self.action_type == 'continuous':
             a = sth.action_index2one_hot(a, self.a_dim_or_list)
         self.data.add(
-            s.astype(np.float32),
-            visual_s.astype(np.float32),
-            a.astype(np.float32),
-            r[:, np.newaxis].astype(np.float32),
-            s_.astype(np.float32),
-            visual_s_.astype(np.float32),
-            done[:, np.newaxis].astype(np.float32)
+            s,
+            visual_s,
+            a,
+            r[:, np.newaxis],
+            s_,
+            visual_s_,
+            done[:, np.newaxis]
         )
 
     def no_op_store(self, s, visual_s, a, r, s_, visual_s_, done):
@@ -136,13 +136,13 @@ class Policy(Base):
             if not self.action_type == 'continuous':
                 a = sth.action_index2one_hot(a, self.a_dim_or_list)
             self.data.add(
-                s.astype(np.float32),
-                visual_s.astype(np.float32),
-                a.astype(np.float32),
-                r[:, np.newaxis].astype(np.float32),
-                s_.astype(np.float32),
-                visual_s_.astype(np.float32),
-                done[:, np.newaxis].astype(np.float32)
+                s,
+                visual_s,
+                a,
+                r[:, np.newaxis],
+                s_,
+                visual_s_,
+                done[:, np.newaxis]
             )
 
     def clear(self):
