+++ gym_wrapper.py
@@ -107,12 +107,12 @@ class gym_envs(object):
             threading.Thread.join(th)
         obs = np.array([threadpool[i].get_result() for i in range(self.n)])
         obs = self._maybe_one_hot(obs)
-        return obs.astype(np.float32)
+        return obs
 
         # if self.obs_type == 'visual':
-        #     return np.array([threadpool[i].get_result()[np.newaxis, :] for i in range(self.n)]).astype(np.float32)
+        #     return np.array([threadpool[i].get_result()[np.newaxis, :] for i in range(self.n)])
         # else:
-        #     return np.array([threadpool[i].get_result() for i in range(self.n)]).astype(np.float32)
+        #     return np.array([threadpool[i].get_result() for i in range(self.n)])
 
     def step(self, actions, scale=True):
         if scale == True:
@@ -140,7 +140,7 @@ class gym_envs(object):
         obs, reward, done, info = [np.array(e) for e in zip(*results)]
         obs = self._maybe_one_hot(obs)
         self.dones_index = np.where(done)[0]
-        return obs.astype(np.float32), reward.astype(np.float32), done, info
+        return obs, reward, done, info
 
     def partial_reset(self):
         threadpool = []
@@ -153,12 +153,12 @@ class gym_envs(object):
             threading.Thread.join(th)
         obs = np.array([threadpool[i].get_result() for i in range(self.dones_index.shape[0])])
         obs = self._maybe_one_hot(obs, is_partial=True)
-        return obs.astype(np.float32)
+        return obs
 
         # if self.obs_type == 'visual':
-        #     return np.array([threadpool[i].get_result()[np.newaxis, :] for i in range(self.dones_index.shape[0])]).astype(np.float32)
+        #     return np.array([threadpool[i].get_result()[np.newaxis, :] for i in range(self.dones_index.shape[0])])
         # else:
-        #     return np.array([threadpool[i].get_result() for i in range(self.dones_index.shape[0])]).astype(np.float32)
+        #     return np.array([threadpool[i].get_result() for i in range(self.dones_index.shape[0])])
     
     def _get_action_normalize_factor(self):
         '''
